cloud-functions/token-generator/requirements.txt
functions-framework==3.*
google-cloud-kms==2.20.0
